chore: refactor sum tokens and add tests
